// apps/web/src/schemas/generated/sessionDto.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * VN Speech Guardian API
 * Speech-to-Text with toxicity detection for Vietnamese
 * OpenAPI spec version: 1.0
 */
import type { SessionDtoEndedAt } from './sessionDtoEndedAt'

export interface SessionDto {
  /** Session id */
  id: string
  /** User id owning this session */
  userId: string
  device?: string
  lang: string
  startedAt: string
  /** @nullable */
  endedAt: SessionDtoEndedAt
}