// apps/web/src/schemas/generated/sessionListResponseDto.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * VN Speech Guardian API
 * Speech-to-Text with toxicity detection for Vietnamese
 * OpenAPI spec version: 1.0
 */
import type { SessionListItemDto } from './sessionListItemDto'

export interface SessionListResponseDto {
  success: boolean
  data: SessionListItemDto
}